Initial reorganization for mail list plugin structure
