+++ src/admin/library/Free/Joomla/Table/Email.php
@@ -10,12 +10,9 @@ namespace Alledia\OSDownloads\Free\Joomla\Table;
 
 defined('_JEXEC') or die();
 
-use Alledia\Framework\Factory;
 use Alledia\Framework\Joomla\Table\Base as BaseTable;
-use Mailchimp\Mailchimp;
+use Alledia\OSDownloads\Free\MailingLists;
 use JEventDispatcher;
-use Exception;
-
 
 class Email extends BaseTable
 {
@@ -27,34 +24,8 @@ class Email extends BaseTable
     public function __construct(&$db)
     {
         parent::__construct('#__osdownloads_emails', 'id', $db);
-    }
 
-    public function addToMailchimpList()
-    {
-        $app    = Factory::getApplication();
-        $params = $app->getParams('com_osdownloads');
-        $apiKey = $params->get("mailchimp_api", 0);
-        $listId = $params->get("list_id", 0);
-
-        if (!empty($this->email)) {
-            $mc = new Mailchimp($apiKey);
-
-            // Check if the email already exists
-            try {
-                $result = $mc->get("lists/{$listId}/members/" . md5(strtolower($this->email)));
-                $result = $result->toArray();
-            } catch (Exception $e) {
-                $result = array('status' => 'unsubscribed');
-            }
-
-            if ($result['status'] === 'unsubscribed') {
-                // The email is not subscribed. Let's subscribe it.
-                $mc->post("lists/{$listId}/members/", array(
-                    'email_address' => $this->email,
-                    'status'        => 'subscribed'
-                ));
-            }
-        }
+        MailingLists::loadObservers($this);
     }
 
     public function store($updateNulls = false)
