+++ src/admin/library/Free/MailingLists.php
@@ -1,0 +1,37 @@
+<?php
+/**
+ * @package    OSDownloads
+ * @contact    www.joomlashack.com, help@joomlashack.com
+ * @copyright  2018 Open Source Training, LLC. All rights reserved
+ * @license    http://www.gnu.org/licenses/gpl.html GNU/GPL
+ *
+ * This file is part of OSDownloads.
+ *
+ * OSDownloads is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * OSDownloads is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with OSDownloads.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+namespace Alledia\OSDownloads\Free;
+
+use Alledia\OSDownloads\Free\MailingList\MailChimp;
+
+defined('_JEXEC') or die();
+
+abstract class MailingLists
+{
+    public static function loadObservers(\JTable $table)
+    {
+        \JFactory::getApplication()->enqueueMessage(__METHOD__);
+        MailChimp::createObserver($table);
+    }
+}
